Fixpoint math moved to Utility
