+++ src/Engine/OurMath.h
@@ -13,38 +13,9 @@
 
 const float pi = static_cast<float>(M_PI);
 
-int64_t fixpoint_mul(int, int);
-int64_t fixpoint_div(int, int);
-
-// These shouldn't compile:
-void fixpoint_mul(float, float) = delete;
-void fixpoint_div(float, float) = delete;
-void fixpoint_mul(double, double) = delete;
-void fixpoint_div(double, double) = delete;
-
-/**
- * @param value                         Fixed-point value.
- * @return                              Corresponding value as a `float`.
- */
-inline float fixpoint_to_float(int value) {
-    return static_cast<float>(value / 65536.0);
-}
-
 int integer_sqrt(int val);
 int GetDiceResult(unsigned int uNumDice, unsigned int uDiceSides);  // idb
 inline int round_to_int(float x) { return (int)floor(x + 0.5f); }
-
-/**
- * Takes a non-fixpoint vector and normalizes it, resulting in a fixpoint vector.
- */
-inline void normalize_to_fixpoint(int *x, int *y, int *z) {
-    extern int integer_sqrt(int val);
-    int denom = *y * *y + *z * *z + *x * *x;
-    int mult = 65536 / (integer_sqrt(denom) | 1);
-    *x *= mult;
-    *y *= mult;
-    *z *= mult;
-}
 
 uint32_t int_get_vector_length(int32_t x, int32_t y, int32_t z);
